requirements.txt
Django==3.2.13
gunicorn==20.0.4
django-filter==21.1
djangorestframework==3.13.1
